requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
flask-migrate
flask-socketio
eventlet
python-dotenv
mysqlclient
# macOS `brew install mysql-client`
cryptography
gunicorn
flasgger
